rename package library
