+++ app/src/main/java/com/filkom/bcc/gridmenu/MainActivity.java
@@ -4,6 +4,9 @@ import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.filkom.bcc.gridmenulibrary.GridMenu;
+import com.filkom.bcc.gridmenulibrary.MenuItem;
+
 import java.util.ArrayList;
 
 public class MainActivity extends AppCompatActivity {
