+++ gridmenulibrary/src/androidTest/java/com/filkom/bcc/gridmenulibrary/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.filkom.bcc.gridmenu;
+package com.filkom.bcc.gridmenulibrary;
 
 import android.content.Context;
