+++ gridmenulibrary/src/main/java/com/filkom/bcc/gridmenulibrary/GridMenuAdapter.java
@@ -1,4 +1,4 @@
-package com.filkom.bcc.gridmenu;
+package com.filkom.bcc.gridmenulibrary;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -9,8 +9,6 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
-
-import com.filkom.bcc.gridmenulibrary.R;
 
 import java.util.List;
