+++ gridmenulibrary/src/main/java/com/filkom/bcc/gridmenulibrary/MenuItem.java
@@ -1,4 +1,4 @@
-package com.filkom.bcc.gridmenu;
+package com.filkom.bcc.gridmenulibrary;
 
 import android.graphics.drawable.Drawable;
